[PyOV] clean up in tests
