+++ src/bindings/python/tests/test_runtime/test_input_node.py
@@ -4,19 +4,17 @@
 
 import os
 
-from ..conftest import model_path
 from openvino.runtime import Input, RTMap
 from openvino._pyopenvino import DescriptorTensor
 import openvino.runtime.opset12 as ops
 
 from openvino import Core, OVAny, Shape, PartialShape, Type
-
-test_net_xml, test_net_bin = model_path()
+from tests.utils.helpers import get_relu_model
 
 
 def test_input_type(device):
     core = Core()
-    model = core.read_model(model=test_net_xml, weights=test_net_bin)
+    model = get_relu_model()
     compiled_model = core.compile_model(model, device)
     net_input = compiled_model.output(0)
     input_node = net_input.get_node().inputs()[0]
@@ -25,7 +23,7 @@ def test_input_type(device):
 
 def test_const_output_docs(device):
     core = Core()
-    model = core.read_model(model=test_net_xml, weights=test_net_bin)
+    model = get_relu_model()
     compiled_model = core.compile_model(model, device)
     net_input = compiled_model.output(0)
     input_node = net_input.get_node().inputs()[0]
@@ -35,7 +33,7 @@ def test_const_output_docs(device):
 
 def test_input_get_index(device):
     core = Core()
-    model = core.read_model(model=test_net_xml, weights=test_net_bin)
+    model = get_relu_model()
     compiled_model = core.compile_model(model, device)
     net_input = compiled_model.output(0)
     input_node = net_input.get_node().inputs()[0]
@@ -44,7 +42,7 @@ def test_input_get_index(device):
 
 def test_input_element_type(device):
     core = Core()
-    model = core.read_model(model=test_net_xml, weights=test_net_bin)
+    model = get_relu_model()
     compiled_model = core.compile_model(model, device)
     net_input = compiled_model.output(0)
     input_node = net_input.get_node().inputs()[0]
@@ -53,7 +51,7 @@ def test_input_element_type(device):
 
 def test_input_get_shape(device):
     core = Core()
-    model = core.read_model(model=test_net_xml, weights=test_net_bin)
+    model = get_relu_model()
     compiled_model = core.compile_model(model, device)
     net_input = compiled_model.output(0)
     input_node = net_input.get_node().inputs()[0]
@@ -62,7 +60,7 @@ def test_input_get_shape(device):
 
 def test_input_get_partial_shape(device):
     core = Core()
-    model = core.read_model(model=test_net_xml, weights=test_net_bin)
+    model = get_relu_model()
     compiled_model = core.compile_model(model, device)
     net_input = compiled_model.output(0)
     input_node = net_input.get_node().inputs()[0]
@@ -72,7 +70,7 @@ def test_input_get_partial_shape(device):
 
 def test_input_get_source_output(device):
     core = Core()
-    model = core.read_model(model=test_net_xml, weights=test_net_bin)
+    model = get_relu_model()
     compiled_model = core.compile_model(model, device)
     net_input = compiled_model.output(0)
     input_node = net_input.get_node().inputs()[0]
@@ -82,7 +80,7 @@ def test_input_get_source_output(device):
 
 def test_input_get_tensor(device):
     core = Core()
-    model = core.read_model(model=test_net_xml, weights=test_net_bin)
+    model = get_relu_model()
     compiled_model = core.compile_model(model, device)
     net_input = compiled_model.output(0)
     input_node = net_input.get_node().inputs()[0]
@@ -92,7 +90,7 @@ def test_input_get_tensor(device):
 
 def test_input_get_rt_info(device):
     core = Core()
-    model = core.read_model(model=test_net_xml, weights=test_net_bin)
+    model = get_relu_model()
     compiled_model = core.compile_model(model, device)
     net_input = compiled_model.output(0)
     input_node = net_input.get_node().inputs()[0]
@@ -102,7 +100,7 @@ def test_input_get_rt_info(device):
 
 def test_input_rt_info(device):
     core = Core()
-    model = core.read_model(model=test_net_xml, weights=test_net_bin)
+    model = get_relu_model()
     compiled_model = core.compile_model(model, device)
     net_input = compiled_model.output(0)
     input_node = net_input.get_node().inputs()[0]
@@ -126,7 +124,7 @@ def test_input_replace_source_output(device):
 
 def test_input_update_rt_info(device):
     core = Core()
-    model = core.read_model(model=test_net_xml, weights=test_net_bin)
+    model = get_relu_model()
     compiled_model = core.compile_model(model, device)
     net_input = compiled_model.output(0)
     input_node = net_input.get_node().inputs()[0]
