+++ src/bindings/python/tests/test_runtime/test_output_node.py
@@ -4,7 +4,6 @@
 
 import os
 
-from ..conftest import model_path
 import openvino.runtime.opset12 as ops
 from openvino import Type
